pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "StreamlitAuth"
version = '0.0.48'
description = 'User login and authentication for Streamlit apps'
readme = 'README.md'
authors = [
    {name = 'Mohammad Khorasani', email='khorasani.mohammad@gmail.com'},
    {name = 'Alex Melesko', email='alex.melesko@msquaredds.com'}
]
license = {file = 'LICENSE'}
requires-python = '>=3.6'
dependencies = [
    "appengine-python-standard >= 1.1.6",
    "argon2-cffi >= 23.1.0",
    "bcrypt >= 3.1.7",
    "crcmod >= 1.7",
    "cryptography >= 41.0.7",
    "extra-streamlit-components >= 0.1.60",
    "google-auth>=2.27.0",
    "google-cloud-kms >= 2.20.0",
    "PyJWT >=2.3.0",
    "PyYAML >= 5.3.1",
    "streamlit >= 1.18.0"
]
keywords = ['Python', 'Streamlit', 'Authentication', 'Components']
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]

[project.urls]
'base repository' = 'https://github.com/mkhorasani/Streamlit-Authenticator'
'test repository' = 'https://github.com/msquaredds/StreamlitAuth'

[tool.setuptools]
packages = ["StreamlitAuth"]